# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.SocialMedia:
    properties:
      handle:
        type: string
      link:
        type: string
      platform:
        type: string
      profile_id:
        type: integer
      social_media_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para gerenciamento de perfis e currículos com IA.
  title: Resume GenAI API
  version: "1.0"
paths:
...
  /socialmedia/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SocialMedia'
            type: array
      summary: Lista todas as redes sociais
      tags:
      - socialmedias
    post:
      consumes:
      - application/json
      parameters:
      - description: Rede social a ser criada
        in: body
        name: socialmedia
        required: true
        schema:
          $ref: '#/definitions/model.SocialMedia'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.SocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Cria uma nova rede social
      tags:
      - socialmedias
  /socialmedias/{id}:
    delete:
      parameters:
      - description: ID da rede social
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Remove uma rede social
      tags:
      - socialmedias
    get:
      parameters:
      - description: ID da rede social
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SocialMedia'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Busca rede social por ID
      tags:
      - socialmedias
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da rede social
        in: path
        name: id
        required: true
        type: integer
      - description: Rede social atualizada
        in: body
        name: socialmedia
        required: true
        schema:
          $ref: '#/definitions/model.SocialMedia'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SocialMedia'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Atualiza uma rede social
      tags:
      - socialmedias
swagger: "2.0"